Jamfile for gui. STDLIBFLAG added to jambase to allow more flexible control of the /MXy flags which determine whether std. libraries are linked statically or dynamically. Second versions of librpc, libsupp, libdbsupp, and libclient are built using the /MDx flag.
Other gui files were modified to use relative paths for include files.
This means that if you want to build the gui from MSVC you will need
to add ...\p4 to your inlcude paths.

The MAXCMD value in jam.h is still an issue. The gui will not successfully
link untill this is increased or a response file is used.

[git-p4: depot-paths = "//depot/": change = 317]